Correcting aspects that should not have been changed
* Fixed some whitespace that was accidentally  changed.
* Removed a debug flag from the parent pom.

+++ openhtmltopdf-pdfbox/src/main/java/com/openhtmltopdf/pdfboxout/PdfBoxImage.java
@@ -112,7 +112,7 @@ public class PdfBoxImage implements FSImage {
 
             if (height == -1 && _intrinsicWidth != 0) {
                 // Use the width ratio to set the height.
-                setHeight = (int) (((float) setWidth / (float) _intrinsicWidth) * _intrinsicHeight);
+                setHeight = (int) ((setWidth / _intrinsicWidth) * _intrinsicHeight);
             } else {
                 setHeight = height;
             }
@@ -121,7 +121,7 @@ public class PdfBoxImage implements FSImage {
 
             if (_intrinsicHeight != 0) {
                 // Use the height ratio to set the width.
-                setWidth = (int) (((float) setHeight / (float) _intrinsicHeight) * _intrinsicWidth);
+                setWidth = (int) ((setHeight / _intrinsicHeight) * _intrinsicWidth);
             } else {
                 setWidth = 0;
             }
@@ -196,7 +196,7 @@ public class PdfBoxImage implements FSImage {
         BufferedImage rimg;
 
         // Create a rotated image buffer and establish the center of rotation 
-		// from the rotation angle.
+        // from the rotation angle.
         switch (angle) {
             case 90:
                 rimg = new BufferedImage(ih, iw, img.getType());
